+++ emu/src/emu_storage.c
@@ -584,12 +584,12 @@ bool readNvsStats(nvs_stats_t* outStats)
  * @brief Read info about each used entry in NVS. Typically, this should be called once with NULL passed for outEntryInfos,
  * to get the value for numEntryInfos, then memory for outEntryInfos should be allocated, then this should be called again
  *
- * @param outStats If not `NULL`, the NVS stats struct will be written to this memory. It must be allocated before calling readAllNvsEntryInfos()
+ * @param outStats If not `NULL`, the NVS stats struct will be written to this memory. It must be allocated before calling `readAllNvsEntryInfos()`
  * @param outEntryInfos A pointer to an array of NVS entry info structs will be written to this memory
  * @param numEntryInfos If outEntryInfos is `NULL`, this will be set to the length of the given key. Otherwise, it is the number of entry infos to read
  * @return true if the entry infos were read, false if they were not
  */
-bool readAllNvsEntryInfos(nvs_stats_t* outStats, nvs_entry_info_t** outEntryInfos, size_t* numEntryInfos)
+bool readAllNvsEntryInfos(nvs_stats_t* outStats, nvs_entry_info_t** outEntryInfos, uint32_t* numEntryInfos)
 {
     // Open the file
     FILE * nvsFile = fopen(NVS_JSON_FILE, "rb");
